requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
PyMySQL>=1.1.0
cryptography>=41.0.0
